# Projet_Examen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portail Scolaire API",
        "description": "Documentation de l'API du portail scolaire",
        "contact": {
            "email": "diabydaj@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bulletins": {
            "get": {
                "tags": [
                    "Bulletins"
                ],
                "summary": "Liste des bulletins",
                "operationId": "d0978bc68981b146f96f77a29027d8fc",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bulletins"
                ],
                "summary": "Générer un nouveau bulletin",
                "operationId": "b245848f5ce0b7d53146ea75f39e7c9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "eleve_id",
                                    "periode"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bulletin généré"
                    }
                }
            }
        },
        "/api/bulletins/{id}": {
            "get": {
                "tags": [
                    "Bulletins"
                ],
                "summary": "Afficher un bulletin spécifique",
                "operationId": "f3dfee80992f2c51eb7603837ff78297",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Bulletin non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bulletins"
                ],
                "summary": "Mettre à jour un bulletin",
                "operationId": "d5bb983914b1008e7c1b54e88f7ef2a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "periode"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bulletin mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bulletins"
                ],
                "summary": "Supprimer un bulletin",
                "operationId": "3bda0f4f8e8c16ead994951714df805a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bulletin supprimé"
                    }
                }
            }
        },
        "/api/eleves/{eleve}/bulletins": {
            "get": {
                "tags": [
                    "Bulletins"
                ],
                "summary": "Liste des bulletins d'un élève",
                "operationId": "3f0ed34193a7a5141baf0ef6904ea25a",
                "parameters": [
                    {
                        "name": "eleve",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}